Use pkg-config in build process.
Remove the hard-coded compiler and linker flags and instead use the ones
that are supplied by pkg-config. This has been tested and built on Mac
OS X Mountain Lion with the homebrew versions of Python and pkgconfig.

See: http://mxcl.github.com/homebrew/ if on Mac OS X.

For Linux and other Unix-like OSes, make sure that the -dev or -devel
packages of lua and Python for your distribution is installed.