Add call with named parameters from lua to python
If a single parameter is passed when calling a python function and that parameter is a table, then numbered elements are passed as regular arguments and named elements as named parameters. This allows, e.g., plt.plot(x, y) and plt.plot{x, y, c='red'} to coexist. The latter is a lua shortcut for passing a single table as argument. When you want to pass a single table as regular argument use func{{1, 2}} or func({1, 2}, nil)